backend/requirements.txt
fastapi
uvicorn
pymupdf
google-generativeai
firebase-admin
python-multipart
python-dotenv
Pillow

